requirements.txt
asgiref==3.4.1
Babel==2.9.1
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.6
django-phonenumber-field==5.2.0
docopt==0.6.2
gunicorn==20.1.0
idna==3.2
phonenumberslite==8.12.31
pipreqs==0.4.10
pytz==2021.1
requests==2.26.0
sqlparse==0.4.2
urllib3==1.26.6
yarg==0.1.9